// packages/core/src/client/ui/components/types/components-types/main-menu-types/CiMainMenuItem.d.ts
import type { CiMainMenuTarget } from "./CiMainMenuTarget";
export type CiMainMenuItem = {
    id: string;
    label: string;
    url?: string;
    icon?: string;
    hidden?: boolean;
    target?: CiMainMenuTarget;
    subMenu?: Record<string, CiMainMenuItem>;
};
//# sourceMappingURL=CiMainMenuItem.d.ts.map